requirements.txt
aiodns==2.0.0
aiofiles==0.5.0
aiohttp==3.6.2
alabaster==0.7.12
aresponses==2.0.0
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.2.0
Babel==2.8.0
beautifulsoup4==4.9.3
brotlipy==0.7.0
cchardet==2.1.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coverage==5.3
cssselect==1.1.0
cycler==0.10.0
Deprecated==1.2.10
docutils==0.16
fake-useragent==0.1.11
feedfinder2==0.0.4
feedparser==6.0.1
idna==2.10
imagesize==1.2.0
iniconfig==1.0.1
jieba3k==0.35.1
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.2.0
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mirakuru==2.3.0
multidict==4.7.6
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.2
packaging==20.4
pamqp==2.3.0
pandas==1.1.3
pika==1.1.0
Pillow==7.2.0
pluggy==0.13.1
polyglot==16.7.4
port-for==0.4
psutil==5.7.2
py==1.9.0
pycares==3.1.1
https://download.lfd.uci.edu/pythonlibs/x2tqcw5k/pycld2-0.41-cp38-cp38-win_amd64.whl; platform_system=='Windows'
pycld2==0.41; platform_system == 'Linux'
pycountry==20.7.3
pycparser==2.20
pyfakefs==4.1.0
Pygments==2.7.1
https://download.lfd.uci.edu/pythonlibs/x2tqcw5k/PyICU-2.5-cp38-cp38-win_amd64.whl; platform_system=='Windows'
PyICU==2.5; platform_system == 'Linux'
pyparsing==2.4.7
pytest==6.1.1
pytest-aiofiles==0.2.0
pytest-asyncio==0.14.0
pytest-rabbitmq==2.0.1
pytest-redis==2.0.0
python-dateutil==2.8.1
python-logging-rabbitmq==2.0.0
pytz==2020.1
PyYAML==5.3.1
rabbitpy==2.0.1
redis==3.5.3
regex==2020.9.27
requests==2.24.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.2.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tinysegmenter==0.3
tldextract==2.2.3
toml==0.10.1
tqdm==4.50.2
urllib3==1.25.10
uvloop==0.14.0; platform_system == 'Linux'
wincertstore==0.2
wrapt==1.12.1
yarl==1.6.0
